refactor: memstorage.go -> storage.go
